pyproject.toml
[project]
name = "corn_classifier"
version = "0.1.0"
description = "Corn desease classification"
readme = "README.md"
requires-python = ">=3.12"
packages = ["corn_classifier"]
dependencies = [
    "mlflow>=3.7.0",
    "onnx>=1.20.0",
    "onnxruntime>=1.23.2",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "pytorch-lightning>=2.6.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]

[project.scripts]
train = "corn_classifier.train:main"
infer = "corn_classifier.infer:main"
export = "corn_classifier.export:main"

[dependency-groups]
dev = [
    "black>=25.11.0",
    "dvc>=3.64.2",
    "flake8>=7.3.0",
    "hydra-core>=1.3.2",
    "isort>=7.0.0",
    "pre-commit>=4.5.0",
    "ruff>=0.14.8",
]
